requirements.txt
requests
numpy
streamlit
pillow